requirements.txt
Pillow==8.4.0
streamlit==0.89.0
